requirements.txt
yfinance
pandas
requests
beautifulsoup4
tabulate
colorama
jinja2
python-dotenv
